Finalizado
Agora é só otimizar.

+++ Vendas OOp em Java/src/pedidos/clientes/Cliente.java
@@ -24,7 +24,7 @@ public class Cliente {
         return uf;
     }
 
-    public float getCodigoClassificacao() {
+    public int getCodigoClassificacao() {
         return classificacao.getCodigo();
     }
